+++ Changelog.md
@@ -3,7 +3,8 @@
 ### Changes:
 * The player will now automatically evade breath attacks
 * Change the player's colour when HP are below 50%
-* Show where the players's blood was spilled for some time
+* Show where the players's blood or the blood of monsters was spilled
+  for some time
 
 ### Fixed bugs:
 * Fixed crash when triggering traps while very burdened (reported by Denis